refs #97
By default, RewriteEngine should be set to Off to ensure that no problem will be encountered before install. If it is set to On and the user tries to install on a subdirectory, the installer may be not reachable (The server issues a 500 error code if the RewriteBase is not correct or absent).

git-svn-id: https://wush.net/svn/wikka/trunk@741 dbc85537-1805-0410-89d4-ae6f2d57612f