refs #230,  comment:10.
The bug was introduced by [628], which forgot the 'format_option' in the compact() function.

At least, this has shown that the strategy used for this tip is
very, very opaque, and may lead, soon or later, to problem.

todo: more comments to explain the tip.

The formatters/wakka.php, is waiting for a variable named $format_option
This variable is passed to the Wakka::Format() method.
The Wakka::Format() method uses compact('text', 'format_option') 
instead of $text and $format_option. This is obscure.
Wakka::IncludeBuffered() then uses extract($vars) to restore
the value of $format_option.


git-svn-id: https://wush.net/svn/wikka/trunk@775 dbc85537-1805-0410-89d4-ae6f2d57612f