refs #38
I made a script to detect all possible constant used in the Wikka but not defined anywhere. It showed that the following constants aren't defined :

 * RFC_MAX_EMAIL_LINE_LENGTH (libs/Compatibilily.lib.php)
 * NO_PAGES_FOUND_MESSAGE (actions/category/category.php)
 * STATUS_NO_PAGES_EDITED (actions/mychanges/mychanges.php)

For RFC_MAX_EMAIL_LINE_LENGTH, I don't know if it should be defined by PHP, a Web search returned no result explaining the use of this constant. For the other 2, I could make the fix.

In comments.xml and recentchanges.xml handlers, just modified VALID_FORMATS in comments to reflect FEED_VALID_FORMATS.


git-svn-id: https://wush.net/svn/wikka/branches/1.3@1643 dbc85537-1805-0410-89d4-ae6f2d57612f